fix: add CommentButton and LikeButton stories, remove InteractionButton and unused stories
