+++ packages/design-system-component-library-yeahyeahyeah/docs/Typography.stories.mdx
@@ -4,8 +4,6 @@ import { Container } from '../components/layout/Container.tsx';
 import { ImageContainer } from '../components/image/ImageContainer.tsx';
 import { Paragraph } from '../components/typography/Paragraph.tsx';
 import { Heading } from '../components/typography/Heading.tsx';
-import { MumbleHeader } from '../components/mumble/MumbleHeader.tsx';
-import { Mumble } from '../components/mumble/Mumble.tsx';
 export const SampleText = 'Lorem ipsum dolor sit amet, consectetur adipiscing elit.';
 
 <Meta title="Guidelines/Typography" />
